backend/requirements.txt
Flask
Flask-Cors
transformers
torch
python-dotenv
google-generativeai
PyJWT
bcrypt